+++ R/table.R
@@ -69,9 +69,10 @@ html_table.xml_node <- function(x, header = NA, trim = TRUE,
   p <- unique(vapply(ncols, sum, integer(1)))
   maxp <- max(p)
 
-  if (length(p) > 1 & (maxp * n != sum(unlist(nrows)))) {
-    # then malformed table is not parsable by rowspan solution
-    if (!fill) {
+  if (length(p) > 1 & maxp * n != sum(unlist(nrows)) &
+      maxp * n != sum(unlist(ncols))) {
+    # then malformed table is not parsable by smart filling solution
+    if (!fill) { # fill must then be specified to allow filling with NAs
       stop("Table has inconsistent number of columns. ",
            "Do you want fill = TRUE?", call. = FALSE)
     }
@@ -80,24 +81,30 @@ html_table.xml_node <- function(x, header = NA, trim = TRUE,
   values <- lapply(cells, html_text, trim = trim)
   out <- matrix(NA_character_, nrow = n, ncol = maxp)
 
+  # fill colspans right with repetition
   for (i in seq_len(n)) {
     row <- values[[i]]
     ncol <- ncols[[i]]
     col <- 1
-    for (j in seq_len(maxp)) {
-      if (j > length(row)) next
-      out[i, col] <- row[[j]]
+    for (j in seq_len(length(ncol))) {
+      out[i, col:(col+ncol[j]-1)] <- row[[j]]
       col <- col + ncol[j]
     }
   }
 
-  # fill rowspans downwards with repetition
+  # fill rowspans down with repetition
   for (i in seq_len(maxp)) {
-    col <- out[, i]
     for (j in seq_len(n)) {
-      rowspan <- nrows[[j]][i]
+      rowspan <- nrows[[j]][i]; colspan <- ncols[[j]][i]
       if (!is.na(rowspan) & (rowspan > 1)) {
-        for (k in seq_len(rowspan-1)) {
+        if (!is.na(colspan) & (colspan > 1)) {
+          # special case of colspan and rowspan in same cell
+          nrows[[j]] <- c(head(nrows[[j]], i),
+                          rep(rowspan, colspan-1),
+                          tail(nrows[[j]], length(rowspan)-(i+1)))
+          rowspan <- nrows[[j]][i]
+        }
+        for (k in seq_len(rowspan - 1)) {
           l <- head(out[j+k, ], i-1)
           r <- tail(out[j+k, ], maxp-i+1)
           out[j + k, ] <- head(c(l, out[j, i], r), maxp)
@@ -123,6 +130,10 @@ html_table.xml_node <- function(x, header = NA, trim = TRUE,
   names(df) <- col_names
   class(df) <- "data.frame"
   attr(df, "row.names") <- .set_row_names(length(df[[1]]))
+
+  if (length(unique(col_names)) < length(col_names)) {
+    warning('At least two columns have the same name')
+  }
 
   df
 }